+++ pkg/backends/internal/options.go
@@ -1,4 +1,4 @@
-package registry
+package internal
 
 import (
 	clientSetV "github.com/hellofresh/kangal/pkg/kubernetes/generated/clientset/versioned"
